feat:
맵 예외처리까지 완료함
그러나 올바른 map 을 넣어도 There are not enough wall 메시지가 출력 되기
때문에 디버깅 해야 함

+++ libft/ft_putchar_fd.c
@@ -6,15 +6,15 @@
 /*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/11 16:56:35 by jayoon            #+#    #+#             */
-/*   Updated: 2022/07/24 23:10:37 by jayoon           ###   ########.fr       */
+/*   Updated: 2022/08/01 12:32:59 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putchar_fd(char c, int fd)
+ssize_t	ft_putchar_fd(char c, int fd)
 {
-	int	len;
+	ssize_t	len;
 
 	len = write(fd, &c, 1);
 	if (len < 0)
