+++ libft/ft_putstr_fd.c
@@ -6,15 +6,15 @@
 /*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/11 17:10:35 by jayoon            #+#    #+#             */
-/*   Updated: 2022/07/27 14:28:32 by jayoon           ###   ########.fr       */
+/*   Updated: 2022/08/01 12:35:57 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putstr_fd(char *s, int fd)
+ssize_t	ft_putstr_fd(char *s, int fd)
 {
-	int	len;
+	ssize_t	len;
 
 	if (!s)
 		return (0);
